\name{acq.comp}
\alias{acq.comp}
\docType{data}
\title{
Acquaintance CAQ Composite
}
\description{
This is the composite of two aquaintace CAQ ratings of a participant's personality from the Riverside Situation Project 
}
\usage{data(acq.comp)}
\format{
  A data frame with 205 observations on the following 100 variables.
   \describe{
    \item{\code{acqCompCAQ001}}{Critical, skeptical, not easily impressed}
    \item{\code{acqCompCAQ002}}{A genuinely dependable and responsible person}
    \item{\code{acqCompCAQ003}}{Has a wide range of interests}
    \item{\code{acqCompCAQ004}}{Talkative}
    \item{\code{acqCompCAQ005}}{Behaves in a giving way toward others}
    \item{\code{acqCompCAQ006}}{Fastidious, perfectionistic}
    \item{\code{acqCompCAQ007}}{Favors conservative values}
    \item{\code{acqCompCAQ008}}{Appears to have a high degree of intellectual capacity}
    \item{\code{acqCompCAQ009}}{Uncomfortable with uncertainty and complexity}
    \item{\code{acqCompCAQ010}}{Anxiety and tension find outlet in bodily symptoms}
    \item{\code{acqCompCAQ011}}{Protective of those close to him or her}
    \item{\code{acqCompCAQ012}}{Tends to be self-defensive}
    \item{\code{acqCompCAQ013}}{Thin-skinned; sensitive to criticism or interpersonal slight}
    \item{\code{acqCompCAQ014}}{Genuinely submissive; accepts domination comfortably}
    \item{\code{acqCompCAQ015}}{Skilled in social techniques of imaginative play, pretending, and humor}
    \item{\code{acqCompCAQ016}}{Introspective and concerned with self as an object}
    \item{\code{acqCompCAQ017}}{Sympathetic and considerate}
    \item{\code{acqCompCAQ018}}{Initiates humor}
    \item{\code{acqCompCAQ019}}{Seeks reassurance from others}
    \item{\code{acqCompCAQ020}}{Has a rapid personal tempo; behaves and acts quickly}
    \item{\code{acqCompCAQ021}}{Arouses nurturant feelings in others}
    \item{\code{acqCompCAQ022}}{Feels a lack of personal meaning in life}
    \item{\code{acqCompCAQ023}}{Extrapunitive; tends to transfer or project blame}
    \item{\code{acqCompCAQ024}}{Prides self on being objective,rational}
    \item{\code{acqCompCAQ025}}{Tends toward over-control of needs and impulses}
    \item{\code{acqCompCAQ026}}{Productive; gets things done}
    \item{\code{acqCompCAQ027}}{Shows condescending behavior in relations with others}
    \item{\code{acqCompCAQ028}}{Tends to arouse liking and acceptance }
    \item{\code{acqCompCAQ029}}{Turned to for advice and reassurance}
    \item{\code{acqCompCAQ030}}{Gives up and withdraws where possible in the face of frustration and adversity}
    \item{\code{acqCompCAQ031}}{Regards self as physically attractive}
    \item{\code{acqCompCAQ032}}{Aware of the impression made on others}
    \item{\code{acqCompCAQ033}}{Calm, relaxed in manner}
    \item{\code{acqCompCAQ034}}{Over-reactive to minor frustrations, irritable}
    \item{\code{acqCompCAQ035}}{Has warmth; has the capacity for close relationships; compassionate}
    \item{\code{acqCompCAQ036}}{Subtly negativistic; tends to undermine and obstruct }
    \item{\code{acqCompCAQ037}}{Guileful and deceitful, manipulative, opportunistic}
    \item{\code{acqCompCAQ038}}{Has hostility toward others}
    \item{\code{acqCompCAQ039}}{Thinks and associates ideas in unusual ways; has unconventional thought processes}
    \item{\code{acqCompCAQ040}}{Vulnerable to real or fancied threat, generally fearful}
    \item{\code{acqCompCAQ041}}{Moralistic}
    \item{\code{acqCompCAQ042}}{Reluctant to commit to any definite course of action; tends to delay or avoid action}
    \item{\code{acqCompCAQ043}}{Facially and/or gesturally expressive}
    \item{\code{acqCompCAQ044}}{Evaluates the motivation of others in interpreting situations}
    \item{\code{acqCompCAQ045}}{Has a brittle ego-defense system; does not cope well under stress or strainr}
    \item{\code{acqCompCAQ046}}{Engages in personal fantasy and daydreams}
    \item{\code{acqCompCAQ047}}{Has a readiness to feel guilt}
    \item{\code{acqCompCAQ048}}{Keeps people at a distance; avoids close interpersonal relationships}
    \item{\code{acqCompCAQ049}}{Basically distrustful of people in general}
    \item{\code{acqCompCAQ050}}{Unpredictable and changeable in behavior and attitudes}
    \item{\code{acqCompCAQ051}}{Genuinely values intellectual and cognitive matters}
    \item{\code{acqCompCAQ052}}{Behaves in an assertive fashion}
    \item{\code{acqCompCAQ053}}{Unable to delay gratification}
    \item{\code{acqCompCAQ054}}{Emphasizes being with others; gregarious}
    \item{\code{acqCompCAQ055}}{Self-defeating}
    \item{\code{acqCompCAQ056}}{Responds to humor}
    \item{\code{acqCompCAQ057}}{Interesting, arresting person}
    \item{\code{acqCompCAQ058}}{Enjoys sensuous experiences (touch, taste, smell, physical contact)}
    \item{\code{acqCompCAQ059}}{Concerned with own body and adequacy of physiological functioning}
    \item{\code{acqCompCAQ060}}{Has insight into own motives and behavior}
    \item{\code{acqCompCAQ061}}{Creates and exploits dependency in people}
    \item{\code{acqCompCAQ062}}{Tends to be rebellious and non-conforming}
    \item{\code{acqCompCAQ063}}{Judges self and other in conventional terms}
    \item{\code{acqCompCAQ064}}{Socially perceptive of a wide range of interpersonal cues}
    \item{\code{acqCompCAQ065}}{Pushes and tries to stretch limits}
    \item{\code{acqCompCAQ066}}{Enjoys esthetic impressions; is esthetically reactive}
    \item{\code{acqCompCAQ067}}{Self-indulgent}
    \item{\code{acqCompCAQ068}}{Basically anxious}
    \item{\code{acqCompCAQ069}}{Sensitive to anything that can be construed as a demand}
    \item{\code{acqCompCAQ070}}{Behaves in an ethically consistent manner}
    \item{\code{acqCompCAQ071}}{Has high aspiration level for self}
    \item{\code{acqCompCAQ072}}{Concerned with own adequacy as a person}
    \item{\code{acqCompCAQ073}}{Tends to perceive many different contexts in sexual terms}
    \item{\code{acqCompCAQ074}}{Subjectively unaware of self-concern; feels satisfied with self}
    \item{\code{acqCompCAQ075}}{Has a clear-cut, internally consistent personality}
    \item{\code{acqCompCAQ076}}{Projects feelings and motivations onto others}
    \item{\code{acqCompCAQ077}}{Appears straightforward, forthright, candid in dealing with others}
    \item{\code{acqCompCAQ078}}{Feels cheated and victimized by life; self-pitying}
    \item{\code{acqCompCAQ079}}{Ruminates and has persistent, preoccupying thoughts}
    \item{\code{acqCompCAQ080}}{Interested in members of the opposite sex}
    \item{\code{acqCompCAQ081}}{Physically attractive; good-looking}
    \item{\code{acqCompCAQ082}}{Has fluctuating moods}
    \item{\code{acqCompCAQ083}}{Able to see to the heart of important problems}
    \item{\code{acqCompCAQ084}}{Cheerful}
    \item{\code{acqCompCAQ085}}{Emphasizes communication through action and non-verbal behavior}
    \item{\code{acqCompCAQ086}}{Repressive and dissociative tendencies; denies unpleasant thoughts and conflicts}
    \item{\code{acqCompCAQ087}}{Interprets basically simple and clear-cut situations in complicated and particularizing ways}
    \item{\code{acqCompCAQ088}}{Personally charming}
    \item{\code{acqCompCAQ089}}{Compares self to others}
    \item{\code{acqCompCAQ090}}{Concerned with philosophical problems}
    \item{\code{acqCompCAQ091}}{Power-oriented; values power in self and others}
    \item{\code{acqCompCAQ092}}{Has social poise and presence; appears socially at ease}
    \item{\code{acqCompCAQ093}}{Behaves in gender-appropriate masculine or feminine style and manner}
    \item{\code{acqCompCAQ094}}{Expresses hostile feelings directly}
    \item{\code{acqCompCAQ095}}{Tends to offer advice}
    \item{\code{acqCompCAQ096}}{Values own independence and autonomy}
    \item{\code{acqCompCAQ097}}{Emotionally bland; has flattened affect}
    \item{\code{acqCompCAQ098}}{Verbally fluent; can express ideas well}
    \item{\code{acqCompCAQ099}}{Self-dramatizing; histrionic}
    \item{\code{acqCompCAQ100}}{Does not vary roles; relates to everyone in the same way}
  }
}
\details{
Subjects are listed as Rows (N=205).
CAQ items (\code{\link{caq.items}}) (100 items) 
}
\source{
\url{http://psy2.fau.edu/~shermanr/index.html}
}
\references{
Sherman, R. A., Nave, C. S., & Funder, D. C. (2010). Situational similarity and personality predict behavioral consistency. Journal of Personality and Social Psychology, 99(2), 330-343
}
\examples{
data(acq.comp)
head(acq.comp) #composites of two ratings
}
\keyword{datasets}
