\name{acq1}
\alias{acq1}
\docType{data}
\title{
Aquaintance (Number 1) CAQ ratings 
}
\description{
This is an aquaintance rating of a participant's personality in the Riverside Situation Project.
}
\usage{data(acq1)}
\format{
  A data frame with 205 observations on the following 100 variables.
 \describe{
    \item{\code{acq1CAQ001}}{Critical, skeptical, not easily impressed}
    \item{\code{acq1CAQ002}}{A genuinely dependable and responsible person}
    \item{\code{acq1CAQ003}}{Has a wide range of interests}
    \item{\code{acq1CAQ004}}{Talkative}
    \item{\code{acq1CAQ005}}{Behaves in a giving way toward others}
    \item{\code{acq1CAQ006}}{Fastidious, perfectionistic}
    \item{\code{acq1CAQ007}}{Favors conservative values}
    \item{\code{acq1CAQ008}}{Appears to have a high degree of intellectual capacity}
    \item{\code{acq1CAQ009}}{Uncomfortable with uncertainty and complexity}
    \item{\code{acq1CAQ010}}{Anxiety and tension find outlet in bodily symptoms}
    \item{\code{acq1CAQ011}}{Protective of those close to him or her}
    \item{\code{acq1CAQ012}}{Tends to be self-defensive}
    \item{\code{acq1CAQ013}}{Thin-skinned; sensitive to criticism or interpersonal slight}
    \item{\code{acq1CAQ014}}{Genuinely submissive; accepts domination comfortably}
    \item{\code{acq1CAQ015}}{Skilled in social techniques of imaginative play, pretending, and humor}
    \item{\code{acq1CAQ016}}{Introspective and concerned with self as an object}
    \item{\code{acq1CAQ017}}{Sympathetic and considerate}
    \item{\code{acq1CAQ018}}{Initiates humor}
    \item{\code{acq1CAQ019}}{Seeks reassurance from others}
    \item{\code{acq1CAQ020}}{Has a rapid personal tempo; behaves and acts quickly}
    \item{\code{acq1CAQ021}}{Arouses nurturant feelings in others}
    \item{\code{acq1CAQ022}}{Feels a lack of personal meaning in life}
    \item{\code{acq1CAQ023}}{Extrapunitive; tends to transfer or project blame}
    \item{\code{acq1CAQ024}}{Prides self on being objective,rational}
    \item{\code{acq1CAQ025}}{Tends toward over-control of needs and impulses}
    \item{\code{acq1CAQ026}}{Productive; gets things done}
    \item{\code{acq1CAQ027}}{Shows condescending behavior in relations with others}
    \item{\code{acq1CAQ028}}{Tends to arouse liking and acceptance }
    \item{\code{acq1CAQ029}}{Turned to for advice and reassurance}
    \item{\code{acq1CAQ030}}{Gives up and withdraws where possible in the face of frustration and adversity}
    \item{\code{acq1CAQ031}}{Regards self as physically attractive}
    \item{\code{acq1CAQ032}}{Aware of the impression made on others}
    \item{\code{acq1CAQ033}}{Calm, relaxed in manner}
    \item{\code{acq1CAQ034}}{Over-reactive to minor frustrations, irritable}
    \item{\code{acq1CAQ035}}{Has warmth; has the capacity for close relationships; compassionate}
    \item{\code{acq1CAQ036}}{Subtly negativistic; tends to undermine and obstruct }
    \item{\code{acq1CAQ037}}{Guileful and deceitful, manipulative, opportunistic}
    \item{\code{acq1CAQ038}}{Has hostility toward others}
    \item{\code{acq1CAQ039}}{Thinks and associates ideas in unusual ways; has unconventional thought processes}
    \item{\code{acq1CAQ040}}{Vulnerable to real or fancied threat, generally fearful}
    \item{\code{acq1CAQ041}}{Moralistic}
    \item{\code{acq1CAQ042}}{Reluctant to commit to any definite course of action; tends to delay or avoid action}
    \item{\code{acq1CAQ043}}{Facially and/or gesturally expressive}
    \item{\code{acq1CAQ044}}{Evaluates the motivation of others in interpreting situations}
    \item{\code{acq1CAQ045}}{Has a brittle ego-defense system; does not cope well under stress or strainr}
    \item{\code{acq1CAQ046}}{Engages in personal fantasy and daydreams}
    \item{\code{acq1CAQ047}}{Has a readiness to feel guilt}
    \item{\code{acq1CAQ048}}{Keeps people at a distance; avoids close interpersonal relationships}
    \item{\code{acq1CAQ049}}{Basically distrustful of people in general}
    \item{\code{acq1CAQ050}}{Unpredictable and changeable in behavior and attitudes}
    \item{\code{acq1CAQ051}}{Genuinely values intellectual and cognitive matters}
    \item{\code{acq1CAQ052}}{Behaves in an assertive fashion}
    \item{\code{acq1CAQ053}}{Unable to delay gratification}
    \item{\code{acq1CAQ054}}{Emphasizes being with others; gregarious}
    \item{\code{acq1CAQ055}}{Self-defeating}
    \item{\code{acq1CAQ056}}{Responds to humor}
    \item{\code{acq1CAQ057}}{Interesting, arresting person}
    \item{\code{acq1CAQ058}}{Enjoys sensuous experiences (touch, taste, smell, physical contact)}
    \item{\code{acq1CAQ059}}{Concerned with own body and adequacy of physiological functioning}
    \item{\code{acq1CAQ060}}{Has insight into own motives and behavior}
    \item{\code{acq1CAQ061}}{Creates and exploits dependency in people}
    \item{\code{acq1CAQ062}}{Tends to be rebellious and non-conforming}
    \item{\code{acq1CAQ063}}{Judges self and other in conventional terms}
    \item{\code{acq1CAQ064}}{Socially perceptive of a wide range of interpersonal cues}
    \item{\code{acq1CAQ065}}{Pushes and tries to stretch limits}
    \item{\code{acq1CAQ066}}{Enjoys esthetic impressions; is esthetically reactive}
    \item{\code{acq1CAQ067}}{Self-indulgent}
    \item{\code{acq1CAQ068}}{Basically anxious}
    \item{\code{acq1CAQ069}}{Sensitive to anything that can be construed as a demand}
    \item{\code{acq1CAQ070}}{Behaves in an ethically consistent manner}
    \item{\code{acq1CAQ071}}{Has high aspiration level for self}
    \item{\code{acq1CAQ072}}{Concerned with own adequacy as a person}
    \item{\code{acq1CAQ073}}{Tends to perceive many different contexts in sexual terms}
    \item{\code{acq1CAQ074}}{Subjectively unaware of self-concern; feels satisfied with self}
    \item{\code{acq1CAQ075}}{Has a clear-cut, internally consistent personality}
    \item{\code{acq1CAQ076}}{Projects feelings and motivations onto others}
    \item{\code{acq1CAQ077}}{Appears straightforward, forthright, candid in dealing with others}
    \item{\code{acq1CAQ078}}{Feels cheated and victimized by life; self-pitying}
    \item{\code{acq1CAQ079}}{Ruminates and has persistent, preoccupying thoughts}
    \item{\code{acq1CAQ080}}{Interested in members of the opposite sex}
    \item{\code{acq1CAQ081}}{Physically attractive; good-looking}
    \item{\code{acq1CAQ082}}{Has fluctuating moods}
    \item{\code{acq1CAQ083}}{Able to see to the heart of important problems}
    \item{\code{acq1CAQ084}}{Cheerful}
    \item{\code{acq1CAQ085}}{Emphasizes communication through action and non-verbal behavior}
    \item{\code{acq1CAQ086}}{Repressive and dissociative tendencies; denies unpleasant thoughts and conflicts}
    \item{\code{acq1CAQ087}}{Interprets basically simple and clear-cut situations in complicated and particularizing ways}
    \item{\code{acq1CAQ088}}{Personally charming}
    \item{\code{acq1CAQ089}}{Compares self to others}
    \item{\code{acq1CAQ090}}{Concerned with philosophical problems}
    \item{\code{acq1CAQ091}}{Power-oriented; values power in self and others}
    \item{\code{acq1CAQ092}}{Has social poise and presence; appears socially at ease}
    \item{\code{acq1CAQ093}}{Behaves in gender-appropriate masculine or feminine style and manner}
    \item{\code{acq1CAQ094}}{Expresses hostile feelings directly}
    \item{\code{acq1CAQ095}}{Tends to offer advice}
    \item{\code{acq1CAQ096}}{Values own independence and autonomy}
    \item{\code{acq1CAQ097}}{Emotionally bland; has flattened affect}
    \item{\code{acq1CAQ098}}{Verbally fluent; can express ideas well}
    \item{\code{acq1CAQ099}}{Self-dramatizing; histrionic}
    \item{\code{acq1CAQ100}}{Does not vary roles; relates to everyone in the same way}
  }
}
\details{
Subjects are listed as Rows (N=205).
CAQ items (\code{\link{caq.items}}) (100 items) 
}
\source{
\url{http://psy2.fau.edu/~shermanr/index.html}
}
\references{
Sherman, R. A., Nave, C. S., & Funder, D. C. (2010). Situational similarity and personality predict behavioral consistency. Journal of Personality and Social Psychology, 99(2), 330-343
}
\examples{
data(acq1)
head(acq1) #aquaintance 1

}
\keyword{datasets}
