\name{acq2}
\alias{acq2}
\docType{data}
\title{
Aquaintance (Number 2) CAQ ratings
}
\description{
This is an aquaintance rating of a participant's personality in the Riverside Situation Project.
}
\usage{data(acq2)}
\format{
  A data frame with 205 observations on the following 100 variables.
   \describe{
    \item{\code{acq2CAQ001}}{Critical, skeptical, not easily impressed}
    \item{\code{acq2CAQ002}}{A genuinely dependable and responsible person}
    \item{\code{acq2CAQ003}}{Has a wide range of interests}
    \item{\code{acq2CAQ004}}{Talkative}
    \item{\code{acq2CAQ005}}{Behaves in a giving way toward others}
    \item{\code{acq2CAQ006}}{Fastidious, perfectionistic}
    \item{\code{acq2CAQ007}}{Favors conservative values}
    \item{\code{acq2CAQ008}}{Appears to have a high degree of intellectual capacity}
    \item{\code{acq2CAQ009}}{Uncomfortable with uncertainty and complexity}
    \item{\code{acq2CAQ010}}{Anxiety and tension find outlet in bodily symptoms}
    \item{\code{acq2CAQ011}}{Protective of those close to him or her}
    \item{\code{acq2CAQ012}}{Tends to be self-defensive}
    \item{\code{acq2CAQ013}}{Thin-skinned; sensitive to criticism or interpersonal slight}
    \item{\code{acq2CAQ014}}{Genuinely submissive; accepts domination comfortably}
    \item{\code{acq2CAQ015}}{Skilled in social techniques of imaginative play, pretending, and humor}
    \item{\code{acq2CAQ016}}{Introspective and concerned with self as an object}
    \item{\code{acq2CAQ017}}{Sympathetic and considerate}
    \item{\code{acq2CAQ018}}{Initiates humor}
    \item{\code{acq2CAQ019}}{Seeks reassurance from others}
    \item{\code{acq2CAQ020}}{Has a rapid personal tempo; behaves and acts quickly}
    \item{\code{acq2CAQ021}}{Arouses nurturant feelings in others}
    \item{\code{acq2CAQ022}}{Feels a lack of personal meaning in life}
    \item{\code{acq2CAQ023}}{Extrapunitive; tends to transfer or project blame}
    \item{\code{acq2CAQ024}}{Prides self on being objective,rational}
    \item{\code{acq2CAQ025}}{Tends toward over-control of needs and impulses}
    \item{\code{acq2CAQ026}}{Productive; gets things done}
    \item{\code{acq2CAQ027}}{Shows condescending behavior in relations with others}
    \item{\code{acq2CAQ028}}{Tends to arouse liking and acceptance }
    \item{\code{acq2CAQ029}}{Turned to for advice and reassurance}
    \item{\code{acq2CAQ030}}{Gives up and withdraws where possible in the face of frustration and adversity}
    \item{\code{acq2CAQ031}}{Regards self as physically attractive}
    \item{\code{acq2CAQ032}}{Aware of the impression made on others}
    \item{\code{acq2CAQ033}}{Calm, relaxed in manner}
    \item{\code{acq2CAQ034}}{Over-reactive to minor frustrations, irritable}
    \item{\code{acq2CAQ035}}{Has warmth; has the capacity for close relationships; compassionate}
    \item{\code{acq2CAQ036}}{Subtly negativistic; tends to undermine and obstruct }
    \item{\code{acq2CAQ037}}{Guileful and deceitful, manipulative, opportunistic}
    \item{\code{acq2CAQ038}}{Has hostility toward others}
    \item{\code{acq2CAQ039}}{Thinks and associates ideas in unusual ways; has unconventional thought processes}
    \item{\code{acq2CAQ040}}{Vulnerable to real or fancied threat, generally fearful}
    \item{\code{acq2CAQ041}}{Moralistic}
    \item{\code{acq2CAQ042}}{Reluctant to commit to any definite course of action; tends to delay or avoid action}
    \item{\code{acq2CAQ043}}{Facially and/or gesturally expressive}
    \item{\code{acq2CAQ044}}{Evaluates the motivation of others in interpreting situations}
    \item{\code{acq2CAQ045}}{Has a brittle ego-defense system; does not cope well under stress or strainr}
    \item{\code{acq2CAQ046}}{Engages in personal fantasy and daydreams}
    \item{\code{acq2CAQ047}}{Has a readiness to feel guilt}
    \item{\code{acq2CAQ048}}{Keeps people at a distance; avoids close interpersonal relationships}
    \item{\code{acq2CAQ049}}{Basically distrustful of people in general}
    \item{\code{acq2CAQ050}}{Unpredictable and changeable in behavior and attitudes}
    \item{\code{acq2CAQ051}}{Genuinely values intellectual and cognitive matters}
    \item{\code{acq2CAQ052}}{Behaves in an assertive fashion}
    \item{\code{acq2CAQ053}}{Unable to delay gratification}
    \item{\code{acq2CAQ054}}{Emphasizes being with others; gregarious}
    \item{\code{acq2CAQ055}}{Self-defeating}
    \item{\code{acq2CAQ056}}{Responds to humor}
    \item{\code{acq2CAQ057}}{Interesting, arresting person}
    \item{\code{acq2CAQ058}}{Enjoys sensuous experiences (touch, taste, smell, physical contact)}
    \item{\code{acq2CAQ059}}{Concerned with own body and adequacy of physiological functioning}
    \item{\code{acq2CAQ060}}{Has insight into own motives and behavior}
    \item{\code{acq2CAQ061}}{Creates and exploits dependency in people}
    \item{\code{acq2CAQ062}}{Tends to be rebellious and non-conforming}
    \item{\code{acq2CAQ063}}{Judges self and other in conventional terms}
    \item{\code{acq2CAQ064}}{Socially perceptive of a wide range of interpersonal cues}
    \item{\code{acq2CAQ065}}{Pushes and tries to stretch limits}
    \item{\code{acq2CAQ066}}{Enjoys esthetic impressions; is esthetically reactive}
    \item{\code{acq2CAQ067}}{Self-indulgent}
    \item{\code{acq2CAQ068}}{Basically anxious}
    \item{\code{acq2CAQ069}}{Sensitive to anything that can be construed as a demand}
    \item{\code{acq2CAQ070}}{Behaves in an ethically consistent manner}
    \item{\code{acq2CAQ071}}{Has high aspiration level for self}
    \item{\code{acq2CAQ072}}{Concerned with own adequacy as a person}
    \item{\code{acq2CAQ073}}{Tends to perceive many different contexts in sexual terms}
    \item{\code{acq2CAQ074}}{Subjectively unaware of self-concern; feels satisfied with self}
    \item{\code{acq2CAQ075}}{Has a clear-cut, internally consistent personality}
    \item{\code{acq2CAQ076}}{Projects feelings and motivations onto others}
    \item{\code{acq2CAQ077}}{Appears straightforward, forthright, candid in dealing with others}
    \item{\code{acq2CAQ078}}{Feels cheated and victimized by life; self-pitying}
    \item{\code{acq2CAQ079}}{Ruminates and has persistent, preoccupying thoughts}
    \item{\code{acq2CAQ080}}{Interested in members of the opposite sex}
    \item{\code{acq2CAQ081}}{Physically attractive; good-looking}
    \item{\code{acq2CAQ082}}{Has fluctuating moods}
    \item{\code{acq2CAQ083}}{Able to see to the heart of important problems}
    \item{\code{acq2CAQ084}}{Cheerful}
    \item{\code{acq2CAQ085}}{Emphasizes communication through action and non-verbal behavior}
    \item{\code{acq2CAQ086}}{Repressive and dissociative tendencies; denies unpleasant thoughts and conflicts}
    \item{\code{acq2CAQ087}}{Interprets basically simple and clear-cut situations in complicated and particularizing ways}
    \item{\code{acq2CAQ088}}{Personally charming}
    \item{\code{acq2CAQ089}}{Compares self to others}
    \item{\code{acq2CAQ090}}{Concerned with philosophical problems}
    \item{\code{acq2CAQ091}}{Power-oriented; values power in self and others}
    \item{\code{acq2CAQ092}}{Has social poise and presence; appears socially at ease}
    \item{\code{acq2CAQ093}}{Behaves in gender-appropriate masculine or feminine style and manner}
    \item{\code{acq2CAQ094}}{Expresses hostile feelings directly}
    \item{\code{acq2CAQ095}}{Tends to offer advice}
    \item{\code{acq2CAQ096}}{Values own independence and autonomy}
    \item{\code{acq2CAQ097}}{Emotionally bland; has flattened affect}
    \item{\code{acq2CAQ098}}{Verbally fluent; can express ideas well}
    \item{\code{acq2CAQ099}}{Self-dramatizing; histrionic}
    \item{\code{acq2CAQ100}}{Does not vary roles; relates to everyone in the same way}
  }
}
\details{
Subjects are listed as Rows (N=205).
CAQ items (\code{\link{caq.items}}) (100 items) 
}
\source{
\url{http://psy2.fau.edu/~shermanr/index.html}
}
\references{
Sherman, R. A., Nave, C. S., & Funder, D. C. (2010). Situational similarity and personality predict behavioral consistency. Journal of Personality and Social Psychology, 99(2), 330-343

}
\examples{
data(acq2)
head(acq2) #aquaintance 2 
}
\keyword{datasets}
