\name{bfi.set}
\alias{bfi.set}
\docType{data}
\title{
Big Five Invetory Set
}
\description{
These are Big Five Inventory self ratings of participants from the Riverside Situation Project.
}
\usage{data(bfi.set)}
\format{
  A data frame with 205 observations on the following 44 variables.
  \describe{
    \item{\code{sBFI1}}{Is talkative} 
    \item{\code{sBFI2}}{Tends to find fault with others}
    \item{\code{sBFI3}}{Does a thorough job }
    \item{\code{sBFI4}}{Is depressed, blue }
    \item{\code{sBFI5}}{Is original, comes up with new ideas }
    \item{\code{sBFI6}}{Is reserved }
    \item{\code{sBFI7}}{Is helpful and unselfish with others }
    \item{\code{sBFI8}}{Can be somewhat careless }
    \item{\code{sBFI9}}{Is relaxed, handles stress well }
    \item{\code{sBFI10}}{Is curious about many different things }
    \item{\code{sBFI11}}{Is full of energy }
    \item{\code{sBFI12}}{Starts quarrels with others }
    \item{\code{sBFI13}}{Is a reliable worker }
    \item{\code{sBFI14}}{Can be tense }
    \item{\code{sBFI15}}{Is ingenious, a deep thinker }
    \item{\code{sBFI16}}{Generates a lot of enthusiasm }
    \item{\code{sBFI17}}{Has a forgiving nature }
    \item{\code{sBFI18}}{Tends to be disorganized }
    \item{\code{sBFI19}}{Worries a lot }
    \item{\code{sBFI20}}{Has an active imagination }
    \item{\code{sBFI21}}{Tends to be quiet }
    \item{\code{sBFI22}}{Is generally trusting }
    \item{\code{sBFI23}}{Tends to be lazy}
    \item{\code{sBFI24}}{Is emotionally stable, not easily upset}
    \item{\code{sBFI25}}{Is inventive}
    \item{\code{sBFI26}}{Has an assertive personality}
    \item{\code{sBFI27}}{Can be cold and aloof}
    \item{\code{sBFI28}}{Perseveres until the task is finished}
    \item{\code{sBFI29}}{Can be moody}
    \item{\code{sBFI30}}{Values artistic, aesthetic experiences}
    \item{\code{sBFI31}}{Is sometimes shy, inhibited}
    \item{\code{sBFI32}}{Is considerate and kind to almost everyone}
    \item{\code{sBFI33}}{Does things efficiently}
    \item{\code{sBFI34}}{Remains calm in tense situations}
    \item{\code{sBFI35}}{Prefers work that is routine}
    \item{\code{sBFI36}}{Is outgoing, sociable}
    \item{\code{sBFI37}}{Is sometimes rude to others}
    \item{\code{sBFI38}}{Makes plans and follows through with them}
    \item{\code{sBFI39}}{Gets nervous easily}
    \item{\code{sBFI40}}{Likes to reflect, play with ideas}
    \item{\code{sBFI41}}{Has few artistic interests}
    \item{\code{sBFI42}}{Likes to cooperate with others}
    \item{\code{sBFI43}}{Is easily distracted}
    \item{\code{sBFI44}}{Is sophisticated in art, music, or literature}
   }
}
\details{
Participants (N=205) are in rows.
BFI items (44 items) are in the columns.
}
\source{
\url{http://psy2.fau.edu/~shermanr/index.html}
}
\references{
Sherman, R. A., Nave, C. S., & Funder, D. C. (2010). Situational similarity and personality predict behavioral consistency. Journal of Personality and Social Psychology, 99(2), 330-343
}

\examples{
data(bfi.set)
head(bfi.set)
}
\keyword{datasets}
