\name{caq.items}
\alias{caq.items}
\docType{data}
\title{
CAQ Items
}
\description{
This is the item content from the California Adult Q-Sort (Block, 1961), a 100 item personality measure.
}
\usage{data(caq.items)}
\format{
  A data frame with content from 100 CAQ items. 
\describe{
	\item{items}{
	{CAQ001 - Critical, skeptical, not easily impressed}
	{\cr CAQ002 - A genuinely dependable and responsible person}
	{\cr CAQ003 - Has a wide range of interests}
	{\cr CAQ004 - Talkative}
	{\cr CAQ005 - Behaves in a giving way toward others}
	{\cr CAQ006 - Fastidious, perfectionistic}
	{\cr CAQ007 - Favors conservative values}
	{\cr CAQ008 - Appears to have a high degree of intellectual capacity}
    {\cr CAQ009 - Uncomfortable with uncertainty and complexity}
    {\cr CAQ010 - Anxiety and tension find outlet in bodily symptoms}
    {\cr CAQ011 - Protective of those close to him or her}
    {\cr CAQ012 - Tends to be self-defensive}
    {\cr CAQ013 - Thin-skinned; sensitive to criticism or interpersonal slight}
    {\cr CAQ014 - Genuinely submissive; accepts domination comfortably}
    {\cr CAQ015 - Skilled in social techniques of imaginative play, pretending, and humor}
    {\cr CAQ016 - Introspective and concerned with self as an object}
    {\cr CAQ017 - Sympathetic and considerate}
    {\cr CAQ018 - Initiates humor}
    {\cr CAQ019 - Seeks reassurance from others}
    {\cr CAQ020 - Has a rapid personal tempo; behaves and acts quickly}
    {\cr CAQ021 - Arouses nurturant feelings in others}
    {\cr CAQ022 - Feels a lack of personal meaning in life}
    {\cr CAQ023 - Extrapunitive; tends to transfer or project blame}
    {\cr CAQ024 - Prides self on being objective,rational}
    {\cr CAQ025 - Tends toward over-control of needs and impulses}
    {\cr CAQ026 - Productive; gets things done}
    {\cr CAQ027 - Shows condescending behavior in relations with others}
    {\cr CAQ028 - Tends to arouse liking and acceptance }
    {\cr CAQ029 - Turned to for advice and reassurance}
    {\cr CAQ030 - Gives up and withdraws where possible in the face of frustration and adversity}
    {\cr CAQ031 - Regards self as physically attractive}
    {\cr CAQ032 - Aware of the impression made on others}
    {\cr CAQ033 - Calm, relaxed in manner}
    {\cr CAQ034 - Over-reactive to minor frustrations, irritable}
    {\cr CAQ035 - Has warmth; has the capacity for close relationships; compassionate}
    {\cr CAQ036 - Subtly negativistic; tends to undermine and obstruct }
    {\cr CAQ037 - Guileful and deceitful, manipulative, opportunistic}
    {\cr CAQ038 - Has hostility toward others}
    {\cr CAQ039 - Thinks and associates ideas in unusual ways; has unconventional thought processes}
    {\cr CAQ040 - Vulnerable to real or fancied threat, generally fearful}
    {\cr CAQ041 - Moralistic}
    {\cr CAQ042 - Reluctant to commit to any definite course of action; tends to delay or avoid action}
    {\cr CAQ043 - Facially and/or gesturally expressive}
    {\cr CAQ044 - Evaluates the motivation of others in interpreting situations}
    {\cr CAQ045 - Has a brittle ego-defense system; does not cope well under stress or strainr}
    {\cr CAQ046 - Engages in personal fantasy and daydreams}
    {\cr CAQ047 - Has a readiness to feel guilt}
    {\cr CAQ048 - Keeps people at a distance; avoids close interpersonal relationships}
    {\cr CAQ049 - Basically distrustful of people in general}
    {\cr CAQ050 - Unpredictable and changeable in behavior and attitudes}
    {\cr CAQ051 - Genuinely values intellectual and cognitive matters}
    {\cr CAQ052 - Behaves in an assertive fashion}
    {\cr CAQ053 - Unable to delay gratification}
    {\cr CAQ054 - Emphasizes being with others; gregarious}
    {\cr CAQ055 - Self-defeating}
    {\cr CAQ056 - Responds to humor}
    {\cr CAQ057 - Interesting, arresting person}
    {\cr CAQ058 - Enjoys sensuous experiences (touch, taste, smell, physical contact)}
    {\cr CAQ059 - Concerned with own body and adequacy of physiological functioning}
    {\cr CAQ060 - Has insight into own motives and behavior}
    {\cr CAQ061 - Creates and exploits dependency in people}
    {\cr CAQ062 - Tends to be rebellious and non-conforming}
    {\cr CAQ063 - Judges self and other in conventional terms}
    {\cr CAQ064 - Socially perceptive of a wide range of interpersonal cues}
    {\cr CAQ065 - Pushes and tries to stretch limits}
    {\cr CAQ066 - Enjoys esthetic impressions; is esthetically reactive}
    {\cr CAQ067 - Self-indulgent}
    {\cr CAQ068 - Basically anxious}
    {\cr CAQ069 - Sensitive to anything that can be construed as a demand}
    {\cr CAQ070 - Behaves in an ethically consistent manner}
    {\cr CAQ071 - Has high aspiration level for self}
    {\cr CAQ072 - Concerned with own adequacy as a person}
    {\cr CAQ073 - Tends to perceive many different contexts in sexual terms}
    {\cr CAQ074 - Subjectively unaware of self-concern; feels satisfied with self}
    {\cr CAQ075 - Has a clear-cut, internally consistent personality}
    {\cr CAQ076 - Projects feelings and motivations onto others}
    {\cr CAQ077 - Appears straightforward, forthright, candid in dealing with others}
    {\cr CAQ078 - Feels cheated and victimized by life; self-pitying}
    {\cr CAQ079 - Ruminates and has persistent, preoccupying thoughts}
    {\cr CAQ080 - Interested in members of the opposite sex}
    {\cr CAQ081 - Physically attractive; good-looking}
    {\cr CAQ082 - Has fluctuating moods}
    {\cr CAQ083 - Able to see to the heart of important problems}
    {\cr CAQ084 - Cheerful}
    {\cr CAQ085 - Emphasizes communication through action and non-verbal behavior}
    {\cr CAQ086 - Repressive and dissociative tendencies; denies unpleasant thoughts and conflicts}
    {\cr CAQ087 - Interprets basically simple and clear-cut situations in complicated and particularizing ways}
    {\cr CAQ088 - Personally charming}
    {\cr CAQ089 - Compares self to others}
    {\cr CAQ090 - Concerned with philosophical problems}
    {\cr CAQ091 - Power-oriented; values power in self and others}
    {\cr CAQ092 - Has social poise and presence; appears socially at ease}
    {\cr CAQ093 - Behaves in gender-appropriate masculine or feminine style and manner}
    {\cr CAQ094 - Expresses hostile feelings directly}
    {\cr CAQ095 - Tends to offer advice}
    {\cr CAQ096 - Values own independence and autonomy}
    {\cr CAQ097 - Emotionally bland; has flattened affect}
    {\cr CAQ098 - Verbally fluent; can express ideas well}
    {\cr CAQ099 - Self-dramatizing; histrionic}
    {\cr CAQ100 - Does not vary roles; relates to everyone in the same way}
    }
  }
}

\source{
\url{http://psy2.fau.edu/~shermanr/index.html}
}
\references{
Block, J. (1961). The Q-Sort Method in Personality Assessment and Psychiatric Research. Springfield, IL: Charles C. Thomas.
}

\examples{
data(caq.items)
caq.items
}
\keyword{datasets}
