\name{caq}
\alias{caq}
\docType{data}
\title{
Californa Adult Q-Set
}
\description{
These are self ratings of personality using the California Adult Q-Set in the Riverside Situation Project.

}
\usage{data(caq)}
\format{
  A data frame with 205 observations on the following 100 variables.
   \describe{
    \item{\code{sCAQ001}}{Critical, skeptical, not easily impressed}
    \item{\code{sCAQ002}}{A genuinely dependable and responsible person}
    \item{\code{sCAQ003}}{Has a wide range of interests}
    \item{\code{sCAQ004}}{Talkative}
    \item{\code{sCAQ005}}{Behaves in a giving way toward others}
    \item{\code{sCAQ006}}{Fastidious, perfectionistic}
    \item{\code{sCAQ007}}{Favors conservative values}
    \item{\code{sCAQ008}}{Appears to have a high degree of intellectual capacity}
    \item{\code{sCAQ009}}{Uncomfortable with uncertainty and complexity}
    \item{\code{sCAQ010}}{Anxiety and tension find outlet in bodily symptoms}
    \item{\code{sCAQ011}}{Protective of those close to him or her}
    \item{\code{sCAQ012}}{Tends to be self-defensive}
    \item{\code{sCAQ013}}{Thin-skinned; sensitive to criticism or interpersonal slight}
    \item{\code{sCAQ014}}{Genuinely submissive; accepts domination comfortably}
    \item{\code{sCAQ015}}{Skilled in social techniques of imaginative play, pretending, and humor}
    \item{\code{sCAQ016}}{Introspective and concerned with self as an object}
    \item{\code{sCAQ017}}{Sympathetic and considerate}
    \item{\code{sCAQ018}}{Initiates humor}
    \item{\code{sCAQ019}}{Seeks reassurance from others}
    \item{\code{sCAQ020}}{Has a rapid personal tempo; behaves and acts quickly}
    \item{\code{sCAQ021}}{Arouses nurturant feelings in others}
    \item{\code{sCAQ022}}{Feels a lack of personal meaning in life}
    \item{\code{sCAQ023}}{Extrapunitive; tends to transfer or project blame}
    \item{\code{sCAQ024}}{Prides self on being objective,rational}
    \item{\code{sCAQ025}}{Tends toward over-control of needs and impulses}
    \item{\code{sCAQ026}}{Productive; gets things done}
    \item{\code{sCAQ027}}{Shows condescending behavior in relations with others}
    \item{\code{sCAQ028}}{Tends to arouse liking and acceptance }
    \item{\code{sCAQ029}}{Turned to for advice and reassurance}
    \item{\code{sCAQ030}}{Gives up and withdraws where possible in the face of frustration and adversity}
    \item{\code{sCAQ031}}{Regards self as physically attractive}
    \item{\code{sCAQ032}}{Aware of the impression made on others}
    \item{\code{sCAQ033}}{Calm, relaxed in manner}
    \item{\code{sCAQ034}}{Over-reactive to minor frustrations, irritable}
    \item{\code{sCAQ035}}{Has warmth; has the capacity for close relationships; compassionate}
    \item{\code{sCAQ036}}{Subtly negativistic; tends to undermine and obstruct }
    \item{\code{sCAQ037}}{Guileful and deceitful, manipulative, opportunistic}
    \item{\code{sCAQ038}}{Has hostility toward others}
    \item{\code{sCAQ039}}{Thinks and associates ideas in unusual ways; has unconventional thought processes}
    \item{\code{sCAQ040}}{Vulnerable to real or fancied threat, generally fearful}
    \item{\code{sCAQ041}}{Moralistic}
    \item{\code{sCAQ042}}{Reluctant to commit to any definite course of action; tends to delay or avoid action}
    \item{\code{sCAQ043}}{Facially and/or gesturally expressive}
    \item{\code{sCAQ044}}{Evaluates the motivation of others in interpreting situations}
    \item{\code{sCAQ045}}{Has a brittle ego-defense system; does not cope well under stress or strainr}
    \item{\code{sCAQ046}}{Engages in personal fantasy and daydreams}
    \item{\code{sCAQ047}}{Has a readiness to feel guilt}
    \item{\code{sCAQ048}}{Keeps people at a distance; avoids close interpersonal relationships}
    \item{\code{sCAQ049}}{Basically distrustful of people in general}
    \item{\code{sCAQ050}}{Unpredictable and changeable in behavior and attitudes}
    \item{\code{sCAQ051}}{Genuinely values intellectual and cognitive matters}
    \item{\code{sCAQ052}}{Behaves in an assertive fashion}
    \item{\code{sCAQ053}}{Unable to delay gratification}
    \item{\code{sCAQ054}}{Emphasizes being with others; gregarious}
    \item{\code{sCAQ055}}{Self-defeating}
    \item{\code{sCAQ056}}{Responds to humor}
    \item{\code{sCAQ057}}{Interesting, arresting person}
    \item{\code{sCAQ058}}{Enjoys sensuous experiences (touch, taste, smell, physical contact)}
    \item{\code{sCAQ059}}{Concerned with own body and adequacy of physiological functioning}
    \item{\code{sCAQ060}}{Has insight into own motives and behavior}
    \item{\code{sCAQ061}}{Creates and exploits dependency in people}
    \item{\code{sCAQ062}}{Tends to be rebellious and non-conforming}
    \item{\code{sCAQ063}}{Judges self and other in conventional terms}
    \item{\code{sCAQ064}}{Socially perceptive of a wide range of interpersonal cues}
    \item{\code{sCAQ065}}{Pushes and tries to stretch limits}
    \item{\code{sCAQ066}}{Enjoys esthetic impressions; is esthetically reactive}
    \item{\code{sCAQ067}}{Self-indulgent}
    \item{\code{sCAQ068}}{Basically anxious}
    \item{\code{sCAQ069}}{Sensitive to anything that can be construed as a demand}
    \item{\code{sCAQ070}}{Behaves in an ethically consistent manner}
    \item{\code{sCAQ071}}{Has high aspiration level for self}
    \item{\code{sCAQ072}}{Concerned with own adequacy as a person}
    \item{\code{sCAQ073}}{Tends to perceive many different contexts in sexual terms}
    \item{\code{sCAQ074}}{Subjectively unaware of self-concern; feels satisfied with self}
    \item{\code{sCAQ075}}{Has a clear-cut, internally consistent personality}
    \item{\code{sCAQ076}}{Projects feelings and motivations onto others}
    \item{\code{sCAQ077}}{Appears straightforward, forthright, candid in dealing with others}
    \item{\code{sCAQ078}}{Feels cheated and victimized by life; self-pitying}
    \item{\code{sCAQ079}}{Ruminates and has persistent, preoccupying thoughts}
    \item{\code{sCAQ080}}{Interested in members of the opposite sex}
    \item{\code{sCAQ081}}{Physically attractive; good-looking}
    \item{\code{sCAQ082}}{Has fluctuating moods}
    \item{\code{sCAQ083}}{Able to see to the heart of important problems}
    \item{\code{sCAQ084}}{Cheerful}
    \item{\code{sCAQ085}}{Emphasizes communication through action and non-verbal behavior}
    \item{\code{sCAQ086}}{Repressive and dissociative tendencies; denies unpleasant thoughts and conflicts}
    \item{\code{sCAQ087}}{Interprets basically simple and clear-cut situations in complicated and particularizing ways}
    \item{\code{sCAQ088}}{Personally charming}
    \item{\code{sCAQ089}}{Compares self to others}
    \item{\code{sCAQ090}}{Concerned with philosophical problems}
    \item{\code{sCAQ091}}{Power-oriented; values power in self and others}
    \item{\code{sCAQ092}}{Has social poise and presence; appears socially at ease}
    \item{\code{sCAQ093}}{Behaves in gender-appropriate masculine or feminine style and manner}
    \item{\code{sCAQ094}}{Expresses hostile feelings directly}
    \item{\code{sCAQ095}}{Tends to offer advice}
    \item{\code{sCAQ096}}{Values own independence and autonomy}
    \item{\code{sCAQ097}}{Emotionally bland; has flattened affect}
    \item{\code{sCAQ098}}{Verbally fluent; can express ideas well}
    \item{\code{sCAQ099}}{Self-dramatizing; histrionic}
    \item{\code{sCAQ100}}{Does not vary roles; relates to everyone in the same way}
  }
}
\details{
Subjects are listed as Rows (N=205).
CAQ items (\code{\link{caq.items}}) (100 items) 
}
\source{
\url{http://psy2.fau.edu/~shermanr/index.html}
}
\references{
Sherman, R. A., Nave, C. S., & Funder, D. C. (2010). Situational similarity and personality predict behavioral consistency. Journal of Personality and Social Psychology, 99(2), 330-343
}
\examples{
data(caq)
head(caq)
data(caq.items)#lets look at the items for the CAQ
caq.items
}

\keyword{datasets}
