\alias{exsitu}
\name{exsitu}
\docType{data}
\title{External Ratings of Situations}

\description{
These are ratings of 10 situations (columns) completed by external raters on 8 characteristics (rows).
}

\usage{data(exsitu)}

\format{
  A matrix containing ratings of 10 situations (columns) on 8 characteristics (rows).
  }

\details{
Situations are the columns (N=10) and characteristics (N=8) are the rows.
}


\references{
Rauthmann, J. F., Sherman, R. A., & Funder, D. C. (forthcoming). Foundations of situation perception: Towards a psychology of how people form impressions of situations. 
  European Journal of Personality.
}

\examples{
data(exsitu)
exsitu
}
\keyword{datasets}