\name{insitu}
\alias{insitu}
\docType{data}
\title{Internal Situation Ratings}

\description{
These are self-ratings of 10 situations (columns) actually experienced by participants (one rater per situation) on 8 characteristics (rows).
}

\usage{data(insitu)}

\format{
  A matrix containing ratings of 10 situations (columns) on 8 characteristics (rows).
   }
   
\details{
Situations are the columns (N=10) and characteristics are the rows.
}


\references{
Rauthmann, J. F., Sherman, R. A., & Funder, D. C. (forthcoming). Foundations of situation perception: Towards a psychology of how people form impressions of situations. 
  European Journal of Personality.
}

\examples{
data(insitu)
insitu
}
\keyword{datasets}