\name{opt.temp}
\alias{opt.temp}
\docType{data}
\title{
Optimum Template
}
\description{
This is a CAQ Template for an optimally adjusted person
}
\usage{data(opt.temp)}
\format{
  Opt.temp contains scores from 1 to 9  on each of 100 \link{caq.items} for an optimally adjusted person.
  }
\details{
A CAQ description of the Optimally Adjusted Person based on 9 independent clinician ratings. The average profile agreement amongst the ratings was r = .78, implying a spearman-brown reliability of .97. 
}

\references{
Block, J. (1961). The Q-Sort Method in Personality Assessment and Psychiatric Research. Springfield, IL: Charles C. Thomas.
}
\examples{
data(opt.temp)
opt.temp
}
\keyword{datasets}
