\name{rate.caq}
\alias{rate.caq}
\docType{data}
\title{
CAQ Rating
}
\description{
This is a dataset of CAQ ratings of participants' personality completed by research assistants using likert type ratings. 
}
\usage{data(rate.caq)}
\format{
  A data frame with 64 observations on the following 100 variables.
  \describe{
\item{\code{CAQ001}}{Critical, skeptical, not easily impressed}
    \item{\code{CAQ002}}{A genuinely dependable and responsible person}
    \item{\code{CAQ003}}{Has a wide range of interests}
    \item{\code{CAQ004}}{Talkative}
    \item{\code{CAQ005}}{Behaves in a giving way toward others}
    \item{\code{CAQ006}}{Fastidious, perfectionistic}
    \item{\code{CAQ007}}{Favors conservative values}
    \item{\code{CAQ008}}{Appears to have a high degree of intellectual capacity}
    \item{\code{CAQ009}}{Uncomfortable with uncertainty and complexity}
    \item{\code{CAQ010}}{Anxiety and tension find outlet in bodily symptoms}
    \item{\code{CAQ011}}{Protective of those close to him or her}
    \item{\code{CAQ012}}{Tends to be self-defensive}
    \item{\code{CAQ013}}{Thin-skinned; sensitive to criticism or interpersonal slight}
    \item{\code{CAQ014}}{Genuinely submissive; accepts domination comfortably}
    \item{\code{CAQ015}}{Skilled in social techniques of imaginative play, pretending, and humor}
    \item{\code{CAQ016}}{Introspective and concerned with self as an object}
    \item{\code{CAQ017}}{Sympathetic and considerate}
    \item{\code{CAQ018}}{Initiates humor}
    \item{\code{CAQ019}}{Seeks reassurance from others}
    \item{\code{CAQ020}}{Has a rapid personal tempo; behaves and acts quickly}
    \item{\code{CAQ021}}{Arouses nurturant feelings in others}
    \item{\code{CAQ022}}{Feels a lack of personal meaning in life}
    \item{\code{CAQ023}}{Extrapunitive; tends to transfer or project blame}
    \item{\code{CAQ024}}{Prides self on being objective,rational}
    \item{\code{CAQ025}}{Tends toward over-control of needs and impulses}
    \item{\code{CAQ026}}{Productive; gets things done}
    \item{\code{CAQ027}}{Shows condescending behavior in relations with others}
    \item{\code{CAQ028}}{Tends to arouse liking and acceptance }
    \item{\code{CAQ029}}{Turned to for advice and reassurance}
    \item{\code{CAQ030}}{Gives up and withdraws where possible in the face of frustration and adversity}
    \item{\code{CAQ031}}{Regards self as physically attractive}
    \item{\code{CAQ032}}{Aware of the impression made on others}
    \item{\code{CAQ033}}{Calm, relaxed in manner}
    \item{\code{CAQ034}}{Over-reactive to minor frustrations, irritable}
    \item{\code{CAQ035}}{Has warmth; has the capacity for close relationships; compassionate}
    \item{\code{CAQ036}}{Subtly negativistic; tends to undermine and obstruct }
    \item{\code{CAQ037}}{Guileful and deceitful, manipulative, opportunistic}
    \item{\code{CAQ038}}{Has hostility toward others}
    \item{\code{CAQ039}}{Thinks and associates ideas in unusual ways; has unconventional thought processes}
    \item{\code{CAQ040}}{Vulnerable to real or fancied threat, generally fearful}
    \item{\code{CAQ041}}{Moralistic}
    \item{\code{CAQ042}}{Reluctant to commit to any definite course of action; tends to delay or avoid action}
    \item{\code{CAQ043}}{Facially and/or gesturally expressive}
    \item{\code{CAQ044}}{Evaluates the motivation of others in interpreting situations}
    \item{\code{CAQ045}}{Has a brittle ego-defense system; does not cope well under stress or strainr}
    \item{\code{CAQ046}}{Engages in personal fantasy and daydreams}
    \item{\code{CAQ047}}{Has a readiness to feel guilt}
    \item{\code{CAQ048}}{Keeps people at a distance; avoids close interpersonal relationships}
    \item{\code{CAQ049}}{Basically distrustful of people in general}
    \item{\code{CAQ050}}{Unpredictable and changeable in behavior and attitudes}
    \item{\code{CAQ051}}{Genuinely values intellectual and cognitive matters}
    \item{\code{CAQ052}}{Behaves in an assertive fashion}
    \item{\code{CAQ053}}{Unable to delay gratification}
    \item{\code{CAQ054}}{Emphasizes being with others; gregarious}
    \item{\code{CAQ055}}{Self-defeating}
    \item{\code{CAQ056}}{Responds to humor}
    \item{\code{CAQ057}}{Interesting, arresting person}
    \item{\code{CAQ058}}{Enjoys sensuous experiences (touch, taste, smell, physical contact)}
    \item{\code{CAQ059}}{Concerned with own body and adequacy of physiological functioning}
    \item{\code{CAQ060}}{Has insight into own motives and behavior}
    \item{\code{CAQ061}}{Creates and exploits dependency in people}
    \item{\code{CAQ062}}{Tends to be rebellious and non-conforming}
    \item{\code{CAQ063}}{Judges self and other in conventional terms}
    \item{\code{CAQ064}}{Socially perceptive of a wide range of interpersonal cues}
    \item{\code{CAQ065}}{Pushes and tries to stretch limits}
    \item{\code{CAQ066}}{Enjoys esthetic impressions; is esthetically reactive}
    \item{\code{CAQ067}}{Self-indulgent}
    \item{\code{CAQ068}}{Basically anxious}
    \item{\code{CAQ069}}{Sensitive to anything that can be construed as a demand}
    \item{\code{CAQ070}}{Behaves in an ethically consistent manner}
    \item{\code{CAQ071}}{Has high aspiration level for self}
    \item{\code{CAQ072}}{Concerned with own adequacy as a person}
    \item{\code{CAQ073}}{Tends to perceive many different contexts in sexual terms}
    \item{\code{CAQ074}}{Subjectively unaware of self-concern; feels satisfied with self}
    \item{\code{CAQ075}}{Has a clear-cut, internally consistent personality}
    \item{\code{CAQ076}}{Projects feelings and motivations onto others}
    \item{\code{CAQ077}}{Appears straightforward, forthright, candid in dealing with others}
    \item{\code{CAQ078}}{Feels cheated and victimized by life; self-pitying}
    \item{\code{CAQ079}}{Ruminates and has persistent, preoccupying thoughts}
    \item{\code{CAQ080}}{Interested in members of the opposite sex}
    \item{\code{CAQ081}}{Physically attractive; good-looking}
    \item{\code{CAQ082}}{Has fluctuating moods}
    \item{\code{CAQ083}}{Able to see to the heart of important problems}
    \item{\code{CAQ084}}{Cheerful}
    \item{\code{CAQ085}}{Emphasizes communication through action and non-verbal behavior}
    \item{\code{CAQ086}}{Repressive and dissociative tendencies; denies unpleasant thoughts and conflicts}
    \item{\code{CAQ087}}{Interprets basically simple and clear-cut situations in complicated and particularizing ways}
    \item{\code{CAQ088}}{Personally charming}
    \item{\code{CAQ089}}{Compares self to others}
    \item{\code{CAQ090}}{Concerned with philosophical problems}
    \item{\code{CAQ091}}{Power-oriented; values power in self and others}
    \item{\code{CAQ092}}{Has social poise and presence; appears socially at ease}
    \item{\code{CAQ093}}{Behaves in gender-appropriate masculine or feminine style and manner}
    \item{\code{CAQ094}}{Expresses hostile feelings directly}
    \item{\code{CAQ095}}{Tends to offer advice}
    \item{\code{CAQ096}}{Values own independence and autonomy}
    \item{\code{CAQ097}}{Emotionally bland; has flattened affect}
    \item{\code{CAQ098}}{Verbally fluent; can express ideas well}
    \item{\code{CAQ099}}{Self-dramatizing; histrionic}
    \item{\code{CAQ100}}{Does not vary roles; relates to everyone in the same way}

  }
}

\source{
\url{http://psy2.fau.edu/~shermanr/index.html}
}
\references{
Serfass, D. G., & Sherman, R. A. (2013). A methodological note on ordered q-sort ratings. Journal of Research in Personality, 47(12), 853-858
}
\examples{
data(rate.caq)
head(rate.caq)
}
\keyword{datasets}
