\name{rbqv3.items}
\alias{rbqv3.items}
\docType{data}
\title{
RBQ Items 
}
\description{
This is the abreivated content from the Riverside Behavioral Q-Sort version 3 (67 items).

}
\usage{data(rbqv3.items)}
\format{
  A data frame with 67 observations on the following variable.
  \describe{
    \item{items}{ {RBQ001 - Interviews others (if present) (e.g., asks a series of questions)}
    {\cr RBQ002 - Volunteers a large amount of information about self}
    {\cr RBQ003 - Seems interested in what someone had to say (Disregard whether interest appears "genuine" or "polite")}
    {\cr RBQ004 - Tries to control the situation (Disregard whether attempts at control succeed or not)}
    {\cr RBQ005 - Dominates the situation (Disregard intention, e.g., if P dominates the situation "by default" because other(s) present do very little, this item should receive high placement)}
    {\cr RBQ006 - Appears to be relaxed and comfortable}
    {\cr RBQ007 - Exhibits social skills (e.g., does things to make other(s) comfortable, keeps conversation moving, entertains or charms other(s))}
    {\cr RBQ008 - Is reserved and unexpressive (e.g., expresses little affect; acts in a stiff, formal manner)}
    {\cr RBQ009 - Laughs frequently (Disregard whether laughter appears to be "nervous" or "genuine")}
    {\cr RBQ010 - Smiles frequently}
    {\cr RBQ011 - Is physically animated; moves around a great deal}
    {\cr RBQ012 - Seems to like other(s) present (e.g., would probably like to be friends with them)}
    {\cr RBQ013 - Exhibits an awkward interpersonal style (e.g., seems to have difficulty knowing what to say, mumbles, fails to respond to other(s)' conversational advances)}
    {\cr RBQ014 - Compares self to other(s) (whether others are present or not)}
    {\cr RBQ015 - Shows high enthusiasm and a high energy level}
    {\cr RBQ016 - Shows a wide range of interests (e.g., talks about many topics)}
    {\cr RBQ017 - Talks at rather than with other(s) (e.g., conducts a monologue, ignores what others say)}
    {\cr RBQ018 - Expresses agreement frequently (High placement implies agreement is expressed unusually often, e.g., in response to each and every statement made. Low placement implies unusual lack of expression of agreement.)}
    {\cr RBQ019 - Expresses criticism (of anybody or anything) (Low placement implies expresses praise)}
    {\cr RBQ020 - Is talkative (as observed in this situation)}
    {\cr RBQ021 - Expresses insecurity (e.g., seems touchy or overly sensitive)}
    {\cr RBQ022 - Shows physical signs of tension or anxiety (e.g., fidgets nervously, voice wavers)(Lack of signs of anxiety = middle placement; low placement = lack of signs under circumstances where you would expect to see them)}
    {\cr RBQ023 - Exhibits a high degree of intelligence (NB: At issue is what is displayed in the interaction not what may or may not be latent. Thus, give this item high placement only if P actually says or does something of high intelligence. Low placement implies exhibition of low intelligence; medium placement = no information one way or the other)}
    {\cr RBQ024 - Expresses sympathy (to anyone, i.e., including conversational references)(Low placement implies unusual lack of sympathy)}
    {\cr RBQ025 - Initiates humor}
    {\cr RBQ026 - Seeks reassurance (e.g., asks for agreement, fishes for praise)}
    {\cr RBQ027 - Exhibits condescending behavior (e.g., acts as if self is superior to others [present, or otherwise])(Low placement implies acting inferior)}
    {\cr RBQ028 - Seems likable (to other(s) present)}
    {\cr RBQ029 - Seeks advice}
    {\cr RBQ030 - Appears to regard self as physically attractive}
    {\cr RBQ031 - Acts irritated}
    {\cr RBQ032 - Expresses warmth (to anyone, e.g., include any references to "my close friend," etc)}
    {\cr RBQ033 - Tries to undermine, sabotage or obstruct}
    {\cr RBQ034 - Expresses hostility (no matter toward whom or what)}
    {\cr RBQ035 - Is unusual or unconventional in appearance}
    {\cr RBQ036 - Behaves in a fearful or timid manner}
    {\cr RBQ037 - Is expressive in face, voice or gestures}
    {\cr RBQ038 - Expresses interest in fantasy or daydreams (Low placement only if such interest is explicitly disavowed)}
    {\cr RBQ039 - Expresses guilt (about anything)}
    {\cr RBQ040 - Keep other(s) at a distance; avoids development of any sort of interpersonal relationship (Low placement implies behavior to get close to other(s))}
    {\cr RBQ041 - Shows interest in intellectual or cognitive matters (e.g., by discussing an intellectual idea in detail or with enthusiasm)}
    {\cr RBQ042 - Seems to enjoy the situation}
    {\cr RBQ043 - Says or does something interesting}
    {\cr RBQ044 - Says negative things about self (e.g., is self-critical; expresses feelings of inadequacy)}
    {\cr RBQ045 - Displays ambition (e.g., passionate discussion of career plans, course grades, opportunities to make money)}
    {\cr RBQ046 - Blames others (for anything)}
    {\cr RBQ047 - Expresses self-pity or feelings of victimization}
    {\cr RBQ048 - Expresses sexual interest (e.g., acts attracted to someone present; expresses interest in dating or sexual matters in general)}
    {\cr RBQ049 - Behaves in a cheerful manner}
    {\cr RBQ050 - Gives up when faced with obstacles (Low placement implies unusual persistence)}
    {\cr RBQ051 - Behaves in a stereotypically masculine/feminine style or manner (Apply the usual stereotypes appropriate to the P's sex. Low placement implies behavior stereotypical of the opposite sex)}
    {\cr RBQ052 - Offers advice}
    {\cr RBQ053 - Speaks fluently and expresses ideas well}
    {\cr RBQ054 - Emphasizes accomplishments of self, family or acquaintances (Low placement = emphasizes failures of these individuals)}
    {\cr RBQ055 - Behaves in a competitive manner (Low placement implies cooperative behavior)}
    {\cr RBQ056 - Speaks in a loud voice}
    {\cr RBQ057 - Speaks sarcastically (e.g., says things (s)he does not mean; makes facetious comments that are not necessarily funny)}
    {\cr RBQ058 - Makes or approaches physical contact with other(s) (Of any sort, including  sitting  unusually close without touching) (Low  placement implies unusual avoidance of physical contact, such as large interpersonal distance)}
    {\cr RBQ059 - Engages in constant eye contact with someone (Low placement implies unusual lack of eye contact)}
    {\cr RBQ060 - Seems detached from the situation}
    {\cr RBQ061 - Speaks quickly (Low placement = speaks slowly)}
    {\cr RBQ062 - Acts playful}
    {\cr RBQ063 - Other(s) seek advice from P}
    {\cr RBQ064 - Concentrates on/works hard at a task (Low placement implies loafing)}
    {\cr RBQ065 - Engages in physical activity (e.g., works up a sweat)(Low placement = almost completely sedentary)}
    {\cr RBQ066 - Acts in a self-indulgent manner (e.g., spending, eating, or drinking)(Low placement implies self-denial)}
    {\cr RBQ067 - Exhibits physical discomfort or pain (High placement = in excess of what seems proportionate; Low placement implies lack of these signs where expected)}
    
    }
  }    
}
\source{
\url{http://psy2.fau.edu/~shermanr/index.html}
}
\references{
Sherman, R. A., Nave, C. S., & Funder, D. C. (2012). Properties of persons and situations related to overall and distinctive personality-behavior congruence. Journal of Research in Personality, 46, 87-101
}
\examples{
data(rbqv3.items)
rbqv3.items
}
\keyword{datasets}