\name{v5rbq}
\alias{v5rbq}
\docType{data}
\title{
Situation 4 RBQ
}
\description{
This is particpants' self ratings of their own behavior using the RBQ in the 4th of 4 situations.
}
\usage{data(v5rbq)}
\format{
  A data frame with 205 observations on the following 67 variables.
    \describe{
    \item{\code{v5rbq001}}{Interviews others (if present) (e.g., asks a series of questions)}
    \item{\code{v5rbq002}}{Volunteers a large amount of information about self}
    \item{\code{v5rbq003}}{Seems interested in what someone had to say (Disregard whether interest appears "genuine" or "polite")}
    \item{\code{v5rbq004}}{Tries to control the situation (Disregard whether attempts at control succeed or not)}
    \item{\code{v5rbq005}}{Dominates the situation (Disregard intention, e.g., if P dominates the situation "by default" because other(s) present do very little, this item should receive high placement)}
    \item{\code{v5rbq006}}{Appears to be relaxed and comfortable}
    \item{\code{v5rbq007}}{Exhibits social skills (e.g., does things to make other(s) comfortable, keeps conversation moving, entertains or charms other(s))}
    \item{\code{v5rbq008}}{Is reserved and unexpressive (e.g., expresses little affect; acts in a stiff, formal manner)}
    \item{\code{v5rbq009}}{Laughs frequently (Disregard whether laughter appears to be "nervous" or "genuine")}
    \item{\code{v5rbq010}}{Smiles frequently}
    \item{\code{v5rbq011}}{Is physically animated; moves around a great deal}
    \item{\code{v5rbq012}}{Seems to like other(s) present (e.g., would probably like to be friends with them)}
    \item{\code{v5rbq013}}{Exhibits an awkward interpersonal style (e.g., seems to have difficulty knowing what to say, mumbles, fails to respond to other(s)' conversational advances)}
    \item{\code{v5rbq014}}{Compares self to other(s) (whether others are present or not)}
    \item{\code{v5rbq015}}{Shows high enthusiasm and a high energy level}
    \item{\code{v5rbq016}}{Shows a wide range of interests (e.g., talks about many topics)}
    \item{\code{v5rbq017}}{Talks at rather than with other(s) (e.g., conducts a monologue, ignores what others say)}
    \item{\code{v5rbq018}}{Expresses agreement frequently (High placement implies agreement is expressed unusually often, e.g., in response to each and every statement made. Low placement implies unusual lack of expression of agreement.)}
    \item{\code{v5rbq019}}{Expresses criticism (of anybody or anything) (Low placement implies expresses praise)}
    \item{\code{v5rbq020}}{Is talkative (as observed in this situation)}
    \item{\code{v5rbq021}}{Expresses insecurity (e.g., seems touchy or overly sensitive)}
    \item{\code{v5rbq022}}{Shows physical signs of tension or anxiety (e.g., fidgets nervously, voice wavers)(Lack of signs of anxiety = middle placement; low placement = lack of signs under circumstances where you would expect to see them)}
    \item{\code{v5rbq023}}{Exhibits a high degree of intelligence (NB: At issue is what is displayed in the interaction not what may or may not be latent. Thus, give this item high placement only if P actually says or does something of high intelligence. Low placement implies exhibition of low intelligence; medium placement = no information one way or the other)}
    \item{\code{v5rbq024}}{Expresses sympathy (to anyone, i.e., including conversational references)(Low placement implies unusual lack of sympathy)}
    \item{\code{v5rbq025}}{Initiates humor}
    \item{\code{v5rbq026}}{Seeks reassurance (e.g., asks for agreement, fishes for praise)}
    \item{\code{v5rbq027}}{Exhibits condescending behavior (e.g., acts as if self is superior to others [present, or otherwise])(Low placement implies acting inferior)}
    \item{\code{v5rbq028}}{Seems likable (to other(s) present)}
    \item{\code{v5rbq029}}{Seeks advice}
    \item{\code{v5rbq030}}{Appears to regard self as physically attractive}
    \item{\code{v5rbq031}}{Acts irritated}
    \item{\code{v5rbq032}}{Expresses warmth (to anyone, e.g., include any references to "my close friend," etc)}
    \item{\code{v5rbq033}}{Tries to undermine, sabotage or obstruct}
    \item{\code{v5rbq034}}{Expresses hostility (no matter toward whom or what)}
    \item{\code{v5rbq035}}{Is unusual or unconventional in appearance}
    \item{\code{v5rbq036}}{Behaves in a fearful or timid manner}
    \item{\code{v5rbq037}}{Is expressive in face, voice or gestures}
    \item{\code{v5rbq038}}{Expresses interest in fantasy or daydreams (Low placement only if such interest is explicitly disavowed)}
    \item{\code{v5rbq039}}{Expresses guilt (about anything)}
    \item{\code{v5rbq040}}{Keep other(s) at a distance; avoids development of any sort of interpersonal relationship (Low placement implies behavior to get close to other(s))}
    \item{\code{v5rbq041}}{Shows interest in intellectual or cognitive matters (e.g., by discussing an intellectual idea in detail or with enthusiasm)}
    \item{\code{v5rbq042}}{Seems to enjoy the situation}
    \item{\code{v5rbq043}}{Says or does something interesting}
    \item{\code{v5rbq044}}{Says negative things about self (e.g., is self-critical; expresses feelings of inadequacy)}
    \item{\code{v5rbq045}}{Displays ambition (e.g., passionate discussion of career plans, course grades, opportunities to make money)}
    \item{\code{v5rbq046}}{Blames others (for anything)}
    \item{\code{v5rbq047}}{Expresses self-pity or feelings of victimization}
    \item{\code{v5rbq048}}{Expresses sexual interest (e.g., acts attracted to someone present; expresses interest in dating or sexual matters in general)}
    \item{\code{v5rbq049}}{Behaves in a cheerful manner}
    \item{\code{v5rbq050}}{Gives up when faced with obstacles (Low placement implies unusual persistence)}
    \item{\code{v5rbq051}}{Behaves in a stereotypically masculine/feminine style or manner (Apply the usual stereotypes appropriate to the P's sex. Low placement implies behavior stereotypical of the opposite sex)}
    \item{\code{v5rbq052}}{Offers advice}
    \item{\code{v5rbq053}}{Speaks fluently and expresses ideas well}
    \item{\code{v5rbq054}}{Emphasizes accomplishments of self, family or acquaintances (Low placement = emphasizes failures of these individuals)}
    \item{\code{v5rbq055}}{Behaves in a competitive manner (Low placement implies cooperative behavior)}
    \item{\code{v5rbq056}}{Speaks in a loud voice}
    \item{\code{v5rbq057}}{Speaks sarcastically (e.g., says things (s)he does not mean; makes facetious comments that are not necessarily funny)}
    \item{\code{v5rbq058}}{Makes or approaches physical contact with other(s) (Of any sort, including  sitting  unusually close without touching) (Low  placement implies unusual avoidance of physical contact, such as large interpersonal distance)}
    \item{\code{v5rbq059}}{Engages in constant eye contact with someone (Low placement implies unusual lack of eye contact)}
    \item{\code{v5rbq060}}{Seems detached from the situation}
    \item{\code{v5rbq061}}{Speaks quickly (Low placement = speaks slowly)}
    \item{\code{v5rbq062}}{Acts playful}
    \item{\code{v5rbq063}}{Other(s) seek advice from P}
    \item{\code{v5rbq064}}{Concentrates on/works hard at a task (Low placement implies loafing)}
    \item{\code{v5rbq065}}{Engages in physical activity (e.g., works up a sweat)(Low placement = almost completely sedentary)}
    \item{\code{v5rbq066}}{Acts in a self-indulgent manner (e.g., spending, eating, or drinking)(Low placement implies self-denial)}
    \item{\code{v5rbq067}}{Exhibits physical discomfort or pain (High placement = in excess of what seems proportionate; Low placement implies lack of these signs where expected)}
  }
  }
\details{
Subjects are listed as Rows (N=205).
RBQ items (\code{\link{rbqv3.items}}) (100 items) 
}
\source{
\url{http://psy2.fau.edu/~shermanr/index.html}
}
\references{
Sherman, R. A., Nave, C. S., & Funder, D. C. (2010). Situational similarity and personality predict behavioral consistency. Journal of Personality and Social Psychology, 99(2), 330-343
}
\examples{
data(v5rbq)	
head(v5rbq)
data(rbqv3.items)#lets look at the RBQ items
rbqv3.items
}

\keyword{datasets}